+++ src/main/java/de/htwsaar/carpool/repository/UserRepository.java
@@ -3,5 +3,5 @@ package de.htwsaar.carpool.repository;
 import de.htwsaar.carpool.model.CarpoolUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<CarpoolUser, Integer> {
+public interface UserRepository extends JpaRepository<CarpoolUser, Long> {
 }
